+++ src/components/PerformanceCard.tsx
@@ -1,67 +1,46 @@
-import { IconButton, Tooltip } from '@chakra-ui/react';
-import styled from '@emotion/styled';
-import { RiFullscreenFill } from 'react-icons/ri';
-
-import { Colors, Fonts, Spaces } from '@/assets/theme';
-
-const Container = styled.div`
-    width: 350px;
-    height: 300px;
-    border-radius: 8px;
-    border: 1px solid ${Colors.border};
-    margin: ${Spaces.large};
-    padding: ${Spaces.large} ${Spaces.medium};
-
-    &:hover {
-        border-color: #319795;
-    }
-`;
-
-const Title = styled.h2`
-    font-size: ${Fonts.sizes.h3};
-    font-weight: ${Fonts.weights.heavy};
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-`;
-
-const Row = styled.div`
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-`;
+import { Tooltip, Flex, Text } from '@chakra-ui/react';
 
 interface PerformanceCardProps {
     title: string;
     children: React.ReactNode;
-    onClickFullScreen: () => void;
+    onClick: () => void;
 }
 
 const PerformanceCard = ({
     title,
     children,
-    onClickFullScreen,
+    onClick,
 }: PerformanceCardProps): JSX.Element => {
     return (
-        <Container>
-            {children}
-            <Row>
+        <Flex
+            flexDirection="column"
+            width={480}
+            height={360}
+            margin="10px"
+            boxShadow="md"
+            padding={2}
+            backgroundColor="white"
+            justifyContent="space-between"
+            rounded="md"
+            border="2px solid transparent"
+            _hover={{ borderColor: 'teal.500', borderWidth: 2 }}
+            onClick={onClick}
+            cursor="pointer"
+        >
+            <div>{children}</div>
+            <Flex justifyContent="space-between" alignItems="flex-end">
                 <Tooltip
                     label={title}
                     fontSize="xs"
                     hasArrow
                     placement="bottom"
                 >
-                    <Title>{title}</Title>
+                    <Text isTruncated fontWeight="medium" fontSize="sm">
+                        {title}
+                    </Text>
                 </Tooltip>
-                <IconButton
-                    marginLeft="1"
-                    aria-label="Toggle Fullscreen Mode"
-                    icon={<RiFullscreenFill />}
-                    onClick={onClickFullScreen}
-                />
-            </Row>
-        </Container>
+            </Flex>
+        </Flex>
     );
 };