+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+-   New design for compute plan performance page (#203)
 -   On compute plan table, the click area for the checkboxes is the full cell (#214)
 -   Updated all dependencies to latest versions (#216)
 -   Improve chart style (#215)